requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
dj-database-url>=2.1,<3.0
django-redis>=5.4,<6.0
django-cors-headers>=4.3,<5.0
celery[redis]>=5.4,<6.0
django-celery-beat>=2.6,<3.0
gunicorn>=22.0,<23.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<11.0
